registration/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-extensions==3.2.3
psycopg==3.2.3
psycopg-binary==3.2.3
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
